feat: Add support for using a custom renderer
When integrating in an app, the CLI display is undesirable. This will
allow us to collect the progress of iterations, so they can be displayed
in a GUI.

Because CLIDashboardRenderer() writes to the console when ::new() is
called, the code has had to be refactored to defer creation until .build()
is called. This meant that instead of delegating the metric assignments
to the already-created dashboard, we instead need to store them and add
them later.