Allow opting out of SQLite's bundling, or all SQLite usage
Users can now choose between the default sqlite-bundled, sqlite,
or disable the default features to skip all SQLite-related code.

I ended up going with 'sqlite' instead of 'sqlite-system', because of
the additive nature of features - if another crate is using rusqlite
and is using the bundled feature already, 'sqlite-system' is a bit
misleading.

This also switches from bundled-full to bundled, which removes a number
of extra features. From what I can tell they weren't being used, but
please correct me if that's not the case.