Use thread-local RNG to generate IDs (#839)
We've been exploring dividing our data set up into multiple batches,
and training those batches in parallel. I noticed that performance did
not scale with core count, and after some digging, found that this was
mainly due to the Mutex being used to generate IDs. With the following
change, training across 16 cores went from 21s to 4.2s.

thread_rng was previously discussed on #703, but I don't believe that
applies here, as this is just used for UUID creation.