Update Cargo.lock workflow to trigger commit checks
A token used to execute an action cannot be reused in another
workflow to prevent infinite chain of workflow execution.

This commit uses a token from a special GitHub app in the tracel
organization so that commit checks (created by the test workflow
which uses the generic GITHUB_TOKEN) can be triggered upon creation
of a Pull Request by this workflow.