burn-wgpu: Move any use of wgpu into a separate module
Currently the burn-wgpu crate is hardcoded to use the wgpu WebGPU
implementaion. It would be nice to be able to use other WebGPU
implementations as new features may land at different times, and there
may be potential performance gains to unlock. So, separate any use of
the wgpu crate into a separate module, in preparation for adding the
ability to target other WebGPU implementaions.

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>