Add a feature to initialize from an existing wgpu adapter/device/queue
This is useful when interacting with other wgpu applications (eg. displaying a burn tensor as a texture in egui). The existing devices are keyed by the wgpu Device ID. Alternatively they could be keyed per adapter which would be more inline with other burn WgpuDevice's (one per adapter), but also there's no real inherent reason to.

This also involves making Queue into an Arc. Alternatively, this could give up ownership of the queue, but it's helpful to be able to synchronize burn operations and custom wgpu operations.

Nb: Even with this feature it's still tricky to say, copy a burn buffer to a wgpu texture, as there's no access to wgpu bindings in the public Burn API. I've got a feature for that locally but not sure what the best way is, might open an issue for it.