fix(onnx-test): mismatch between expected and actual output types (#3370)
See: https://github.com/tracel-ai/burn/issues/3370#issuecomment-3080046479

The root cause is that the WGPU backend uses `i32` and `u32` to represent integer and boolean values, respectively. However, the expected values in tests were defined using `i64` and `bool`, which align with the default types used by the NdArray backend.

To fix this, features such as `int-i32` and `bool-u32` are introduced to allow switching the expected types depending on the backend.

### Considered alternatives:
1. **Implementing `From` for `TensorData`**: This would reduce boilerplate but obscure the actual data types, which ONNX test contributors are expected to be aware of.
2. **Using helper functions to convert values**: Given that both backend type and expected values vary across tests, such helpers would have limited use and add unnecessary complexity. Furthermore, they make tests less self-contained and harder to reason about locally.